requirements.txt
streamlit==1.36.0
plotly==5.23.0
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
